refactor: add custom error to Authentication use cases
