+++ src/application/errors/authentication/account-not-found-error.ts
@@ -1,0 +1,7 @@
+import { AuthenticationError } from '@/domain/errors'
+
+export class AccountNotFoundError extends AuthenticationError {
+  constructor() {
+    super('Account not found')
+  }
+}
