+++ src/application/errors/authentication/expired-token-error.ts
@@ -1,0 +1,7 @@
+import { AuthenticationError } from '@/domain/errors'
+
+export class ExpiredTokenError extends AuthenticationError {
+  constructor() {
+    super('Expired token')
+  }
+}
