+++ src/application/errors/authentication/index.ts
@@ -1,2 +1,3 @@
+export * from './account-not-found-error'
 export * from './expired-token-error'
 export * from './invalid-token-error'
