+++ src/application/errors/authentication/invalid-token-error.ts
@@ -1,0 +1,7 @@
+import { AuthenticationError } from '@/domain/errors'
+
+export class InvalidTokenError extends AuthenticationError {
+  constructor() {
+    super('Invalid token')
+  }
+}
