--- src/application/errors/expired-token-error.ts
@@ -1,7 +1,0 @@
-import { TokenError } from '@/domain/errors/token-error'
-
-export class ExpiredTokenError extends TokenError {
-  constructor() {
-    super('Expired token')
-  }
-}
