--- src/application/errors/invalid-token-error.ts
@@ -1,7 +1,0 @@
-import { TokenError } from '@/domain/errors/token-error'
-
-export class InvalidTokenError extends TokenError {
-  constructor() {
-    super('Invalid token')
-  }
-}
