+++ src/application/protocols/db/static/authentication/get-account-by-token-repository.ts
@@ -1,3 +1,9 @@
 export interface GetAccountByTokenRepository {
-  getByToken: (token: string, role?: string) => Promise<string | null>
+  getByToken: (token: string, role?: string) => Promise<GetAccountByTokenRepository.Output>
+}
+
+export namespace GetAccountByTokenRepository {
+  export interface Output {
+    id: string
+  }
 }
