+++ src/domain/errors/authentication-error.ts
@@ -1,4 +1,4 @@
-export class TokenError extends Error {
+export class AuthenticationError extends Error {
   constructor(errorMessage: string) {
     super(errorMessage)
     this.name = this.constructor.name
