+++ src/domain/usecases/authentication/get-account-by-token.ts
@@ -1,3 +1,9 @@
 export interface GetAccountByToken {
-  getByToken: (accessToken: string, role?: string) => Promise<string | null>
+  getByToken: (accessToken: string, role?: string) => Promise<GetAccountByToken.Output>
+}
+
+export namespace GetAccountByToken {
+  export interface Output {
+    id: string
+  }
 }
